+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 **BREAKING** Update `imxrt-usbd` to 0.3. Users are required to update to
 `usb-device` 0.3 and its compatible dependencies.
 
+**BREAKING** Remove deprecated APIs
+
+- `clko2::Selection::TracClk`
+- `timer::*PitChan`
+- `lpspi::Disabled::{set_mode, set_watermark}`
+
 ## [0.5.6] 2024-07-05
 
 - Add LPSPI low-level clock configuration APIs.